Don't hash the password before submitting it.
This does not work with our production directory. When you do an LDAP
replace operation, it accepts a plain text password and then hashes it
itself. If you use add, it does not hash it automatically. We'll lean on
this automatic behavior to simplify our code, and I'll remove all the
extra crypto stuff from the codebase.