Update gems and fix some spec organization
Some of the organization of the specs in there is a little wonky, and something
updated and broke one of the admittedly brittle Directory specs. I made some of
the prep around them better (before :each blocks and some convenience methods),
but it's still not great. I'll probably never fix this because there isn't much
reason to. :)